Add files via upload
This upgrade now asks the teacher for the name of the class/exam, the teachers name, the date of the class/exam and the date the register is printed. It prints all student names with student ID's and prints the register to a local printer (must be on the windows platform) as well as displaying the results in a .txt file which can then be shared or stored.